made MindBlow wayyy faster
More threads. More responsive.
The IBFListener calls don't wait for the WinForms updates to finish
now. They put changes onto a few variables using a lock and then
activate a dormant thread to actually run those on the worker thread.
Once it's done, that dormant thread goes back to waiting for an
AutoResetEvent to fire that tells it when there's actually work to
do. So it shouldn't chew up any CPU. I found out about this object
working on this project and now I want to find everywhere in the game
that uses a while loop to wait for stuff, and replace it with this.