Dependency Injection. (+ misc because I'm lazy)
Tired of design patterns ruining your day? Have some more to
patch up the existing ones.
The new system which gets rid of true global state (allowing for
multiple servers in one process for... some reason...) also
adds what I find to be a lot of unnecessary mental strain into
navigating the project because every method in the new areas
takes a Backend argument and gets the rest of the components of
the engine through some weird generic function call.
Even though the introduction of this system is meant to adhere
to OOP Design Patterns (I still hold that "patterns" belong
nowhere near programming, but C# is java-lite, you don't have a
choice) - I had a couple nights to sleep on it and realise that
it actually violates the Hollywood Principle, which is the root
of my problem with it, and that can be alleviated using Dependency
Injection. I'm sure I can dig up more buzzwords if you want.
The new dependency injection framework will fill in fields on a
new object which have the [Dependency] attribute, from the backend
component list. This means that instead of having to go fetch the
engine parts, classes can just make room for them, set the right
attributes, and the framework fills them in. Less tedious.

Speaking of dependencies, I removed one on Discord's desktop client
that was making the game crash on startup. I use Discord Web, the
desktop client freezes on notifications for me. :picardia:

Also I once again cleaned up Mike's horrific approach to threading,
which he said he got from Barnacules.